my first commit in april 2024

+++ remove_product_from_watch_list.php
@@ -1,9 +1,10 @@
 <?php
 session_start();
-require_once  'C:\wamp64\www\lm\conn.php';
+require_once  'conn.php';
 
-
-   $sql = "DELETE FROM `wishlist` WHERE `inventory_item_id` = ".$_POST['remove'];
+   $sql = "DELETE FROM `wishlist` WHERE `inventory_item_id` = ".$_POST['remove']."and  `customer_id` = ".$_SESSION['uid'];
+//$sql = "DELETE FROM `wishlist` WHERE `inventory_item_id` = 8 and  `customer_id` = ".$_SESSION['uid'];
+ //echo $sql;
    $result = $mysqli->query($sql);
    if($result){
    $arr=array('msg' => 'Success');
